+++ frontend/src/ProfilePage.jsx
@@ -4,7 +4,6 @@ import "./css/ProfilePage.css";
 import { Link, useNavigate } from "react-router-dom";
 import { useAuth } from "./contexts/AuthContext";
 import { API_BASE_URL } from "./apiConfig";
-import { FaEye, FaEyeSlash } from "react-icons/fa";
 
 const ProfilePage = ({ t }) => {
   const { currentUser, token, updateUserAvatar } = useAuth();
@@ -20,22 +19,6 @@ const ProfilePage = ({ t }) => {
   const [editing, setEditing] = useState(false);
   const [error, setError] = useState("");
   const [isLoadingAvatar, setIsLoadingAvatar] = useState(false);
-
-  // ✨ state cho đổi mật khẩu
-  const [passwordData, setPasswordData] = useState({
-    oldPassword: "",
-    newPassword: "",
-    confirmPassword: "",
-  });
-  const [passwordError, setPasswordError] = useState("");
-  const [passwordSuccess, setPasswordSuccess] = useState("");
-
-  // ✨ state ẩn/hiện mật khẩu
-  const [showPassword, setShowPassword] = useState({
-    old: false,
-    new: false,
-    confirm: false,
-  });
 
   const navigate = useNavigate();
   const fileInputRef = useRef(null);
@@ -193,36 +176,6 @@ const ProfilePage = ({ t }) => {
     }
   };
 
-  const handlePasswordChange = (e) => {
-    setPasswordData((prev) => ({ ...prev, [e.target.name]: e.target.value }));
-  };
-
-  const handleChangePassword = async () => {
-    if (passwordData.newPassword !== passwordData.confirmPassword) {
-      setPasswordError("Mật khẩu mới và xác nhận không khớp.");
-      return;
-    }
-    try {
-      const config = { headers: { Authorization: `Bearer ${token}` } };
-      await axios.put(
-        `${API_BASE_URL}/api/accounts/${currentUser.accountId}/change-password`,
-        {
-          oldPassword: passwordData.oldPassword,
-          newPassword: passwordData.newPassword,
-        },
-        config
-      );
-      setPasswordSuccess("Đổi mật khẩu thành công!");
-      setPasswordError("");
-      setPasswordData({ oldPassword: "", newPassword: "", confirmPassword: "" });
-    } catch (err) {
-      setPasswordError(
-        "Đổi mật khẩu thất bại: " + (err.response?.data || err.message)
-      );
-      setPasswordSuccess("");
-    }
-  };
-
   if (error) return <div className="profile-error">{error}</div>;
   if (!account || !user) {
     return (
@@ -321,81 +274,6 @@ const ProfilePage = ({ t }) => {
           </button>
         </div>
       )}
-
-      {/* ✨ Form đổi mật khẩu */}
-      <h3 className="profile-subtitle">Đổi mật khẩu</h3>
-
-      {/* Mật khẩu cũ */}
-      <div className="profile-field password-field">
-        <label>Mật khẩu cũ:</label>
-        <div className="password-input-wrapper">
-          <input
-            type={showPassword.old ? "text" : "password"}
-            name="oldPassword"
-            value={passwordData.oldPassword}
-            onChange={handlePasswordChange}
-            className="profile-input"
-          />
-          <span
-            className="password-toggle"
-            onClick={() =>
-              setShowPassword((prev) => ({ ...prev, old: !prev.old }))
-            }
-          >
-            {showPassword.old ? <FaEyeSlash /> : <FaEye />}
-          </span>
-        </div>
-      </div>
-
-      {/* Mật khẩu mới */}
-      <div className="profile-field password-field">
-        <label>Mật khẩu mới:</label>
-        <div className="password-input-wrapper">
-          <input
-            type={showPassword.new ? "text" : "password"}
-            name="newPassword"
-            value={passwordData.newPassword}
-            onChange={handlePasswordChange}
-            className="profile-input"
-          />
-          <span
-            className="password-toggle"
-            onClick={() =>
-              setShowPassword((prev) => ({ ...prev, new: !prev.new }))
-            }
-          >
-            {showPassword.new ? <FaEyeSlash /> : <FaEye />}
-          </span>
-        </div>
-      </div>
-
-      {/* Xác nhận mật khẩu mới */}
-      <div className="profile-field password-field">
-        <label>Xác nhận mật khẩu mới:</label>
-        <div className="password-input-wrapper">
-          <input
-            type={showPassword.confirm ? "text" : "password"}
-            name="confirmPassword"
-            value={passwordData.confirmPassword}
-            onChange={handlePasswordChange}
-            className="profile-input"
-          />
-          <span
-            className="password-toggle"
-            onClick={() =>
-              setShowPassword((prev) => ({ ...prev, confirm: !prev.confirm }))
-            }
-          >
-            {showPassword.confirm ? <FaEyeSlash /> : <FaEye />}
-          </span>
-        </div>
-      </div>
-
-      <button onClick={handleChangePassword} className="profile-button edit-button">
-        Đổi mật khẩu
-      </button>
-      {passwordError && <div className="error-text">{passwordError}</div>}
-      {passwordSuccess && <div className="success-text">{passwordSuccess}</div>}
     </div>
   );
 };
